# libs/shared/src/lib/models/adoption-entity.ts
/* tslint:disable */
/* eslint-disable */
export interface AdoptionEntity {
  adoptorLocation: string;
  adoptorName: string;
  adoptorPhone: string;
  adoptor_email: string;
  createTime: string;
  id: number;
}